refactor(nat): move the MCP export to xr_ai_nat.mcp
Relocate the generic native-function → MCP publisher from
`xr_ai_nat.adapters.mcp` to a top-level `xr_ai_nat.mcp` package
(`create_mcp_server` re-exported from `xr_ai_nat.mcp`). `mcp/` now owns
exposing native capabilities to MCP-only agents; `adapters/` is left for
framework adapters. Importers (transcript-mcp, vlm-mcp, tests) and the
README point at the new path.

Behaviour-preserving: `create_mcp_server`'s signature and behaviour are
unchanged; only the module location moved. No dependency changes.
