+++ agent-mcp-servers/vlm-mcp/vlm_mcp_server/__main__.py
@@ -59,7 +59,7 @@ from loguru import logger
 from nat.builder.workflow_builder import WorkflowBuilder
 
 from xr_ai_logging import setup_logging
-from xr_ai_nat.adapters.mcp import create_mcp_server
+from xr_ai_nat.mcp import create_mcp_server
 from xr_ai_nat.functions.vision import VisionFunctionsConfig
 from xr_ai_models import (
     ModelsConfig,
