+++ docs/changelog.md
@@ -9,6 +9,14 @@ Significant decisions, in reverse-chronological order. Update this whenever a
 non-trivial architectural or design decision is made so the rationale is
 preserved and not re-litigated.
 
+### 2026-07-27 — MCP export lives under `xr_ai_nat.mcp`
+
+The generic native-function → MCP publisher moved from
+`xr_ai_nat.adapters.mcp` to a top-level `xr_ai_nat.mcp` package
+(`xr_ai_nat.mcp.create_mcp_server`). `mcp/` now owns exposing native
+capabilities to MCP-only agents, leaving `adapters/` for framework
+adapters. Public signature is unchanged; only the import path moved.
+
 ### 2026-07-27 — Shared service transport and value models under `xr-ai-nat`
 
 The private correlated msgpack/ZMQ transport shared by service-backed functions
